Move statsbeat code to the proper class and and ensure it's initialized.
