+++ src/declarations/constants.ts
@@ -2,21 +2,3 @@ export const APPLICATION_INSIGHTS_SDK_VERSION = "3.0.0-beta.0";
 export const DEFAULT_BREEZE_ENDPOINT = "https://dc.services.visualstudio.com";
 export const DEFAULT_LIVEMETRICS_ENDPOINT = "https://rt.services.visualstudio.com";
 export const DEFAULT_LIVEMETRICS_HOST = "rt.services.visualstudio.com";
-export const NON_EU_CONNECTION_STRING =
-  "InstrumentationKey=c4a29126-a7cb-47e5-b348-11414998b11e;IngestionEndpoint=https://westus-0.in.applicationinsights.azure.com";
-export const EU_CONNECTION_STRING =
-  "InstrumentationKey=7dc56bab-3c0c-4e9f-9ebb-d1acadee8d0f;IngestionEndpoint=https://westeurope-5.in.applicationinsights.azure.com";
-export const EU_ENDPOINTS = [
-  "westeurope",
-  "northeurope",
-  "francecentral",
-  "francesouth",
-  "germanywestcentral",
-  "norwayeast",
-  "norwaywest",
-  "swedencentral",
-  "switzerlandnorth",
-  "switzerlandwest",
-  "uksouth",
-  "ukwest"
-];
